Move CorporaTestCaseUsingMockAWS to tests.unit.backend.fixtures
