+++ tests/unit/backend/chalice/api_server/test_v1_dataset.py
@@ -6,9 +6,9 @@ import sys
 from furl import furl
 
 from tests.unit.backend.chalice.api_server import BaseAPITest
-from tests.unit.backend.corpora import CorporaTestCaseUsingMockAWS
 from tests.unit.backend.chalice.api_server.mock_auth import MockOauthServer, get_auth_token
 from tests.unit.backend.chalice.api_server.generate_data_mixin import GenerateDataMixin
+from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
 
 
 class TestDataset(BaseAPITest, GenerateDataMixin, CorporaTestCaseUsingMockAWS):
